fix: centralise and correct registration validations
