+++ frontend/src/views/InscriptionUtilisateur.vue
@@ -94,6 +94,7 @@ import {
   DsfrButton,
   DsfrNotice
 } from '@gouvminint/vue-dsfr'
+import useInscriptionValidation from '@/composables/use-inscription-validation'
 import { inscrireUtilisateur } from '@/services/apiService'
 
 const filAriane = [
@@ -109,71 +110,34 @@ const form = reactive({
   acceptTerms: false
 })
 
-// Erreurs champ par champ
-type FormErrors = {
-  nom_utilisateur?: string
-  email?: string
-  password?: string
-  confirmPassword?: string
-  acceptTerms?: string
-}
-
-const errors = reactive<FormErrors>({
-  nom_utilisateur: undefined,
-  email: undefined,
-  password: undefined,
-  confirmPassword: undefined,
-  acceptTerms: undefined
-})
+// Validation centralisée du formulaire d'inscription
+const {
+  errors,
+  validateUsername,
+  validateEmail,
+  validatePassword,
+  validateConfirmPassword,
+  validateAcceptTerms,
+  validateAll,
+  isValid
+} = useInscriptionValidation(form)
 
 // Erreur globale (submit)
 const erreurGlobal = ref<{titre:string,texte:string}|null>(null)
 const isFormulaireSoumis = ref(false)
 const etapeCourante = ref(1)
 
-/********* Validation des champs *********/
-function validateUsername() {
-  errors.nom_utilisateur = /^[\w-]{3,}$/.test(form.nom_utilisateur)
-    ? undefined
-    : 'Pseudonyme invalide (min 3 caractères, lettres chiffres tirets)'
-}
-
-function validateEmail() {
-  errors.email = /^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(form.email)
-    ? undefined
-    : 'Format d’email invalide'
-}
-
-function validatePassword() {
-  errors.password = /(?=.{8,})(?=.*[A-Z])(?=.*[a-z])(?=.*\d)/.test(form.password)
-    ? undefined
-    : 'Mot de passe faible'
-}
-
-function validateConfirmPassword() {
-  errors.confirmPassword = form.confirmPassword === form.password
-    ? undefined
-    : 'Les mots de passe ne correspondent pas'
-}
-
-function validateAcceptTerms() {
-  errors.acceptTerms = form.acceptTerms
-    ? undefined
-    : 'Vous devez accepter les CGU'
-}
-
 // Navigation étapes
 function etapeSuivante() {
-  validateUsername()
-  validateEmail()
-  validatePassword()
-  validateConfirmPassword()
-  validateAcceptTerms()
-  if (Object.values(errors).every(e => e === null)) {
+  validateAll()
+  if (isValid()) {
     erreurGlobal.value = null
     etapeCourante.value = 2
   } else {
-    erreurGlobal.value = { titre: 'Formulaire invalide', texte: 'Veuillez corriger les erreurs' }
+    erreurGlobal.value = {
+      titre: 'Formulaire invalide',
+      texte: 'Veuillez corriger les erreurs'
+    }
   }
 }
 
@@ -184,8 +148,11 @@ function etapePrecedente() {
 
 // Soumission
 async function soumettreFormulaire() {
-  etapeSuivante()
-  if (etapeCourante.value !== 2 || erreurGlobal.value) return
+  validateAll()
+  if (!isValid()) {
+    erreurGlobal.value = { titre: 'Formulaire invalide', texte: 'Veuillez corriger les erreurs' }
+    return
+  }
 
   try {
     await inscrireUtilisateur({
